+++ projects/ngx-single-select/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 8.0.2
+- Updated Component Options
+
 ## 8.0.1
 - Added Keyboard Navigation
 - Added Search for listed items